+++ NEWS.md
@@ -8,6 +8,7 @@ Listing news on any major breaking changes in DFG.  For regular changes, see int
 - Towards distributions serialized via JSON, getting away from custom strings.
 - `LocalDFG` replaces `DefaultDFG` (#844).
 - Optimized creation of CGDFG / `createDfgSessionIfNotExist` (#839, #815).
+- `plotDFG` replaces `dfgplot` (#841, #844).
 
 # v0.16.0